+++ src/provider/client/types/was/index.ts
@@ -1,2 +1,3 @@
 export * from './listWebApps';
 export * from './listWebAppFindings';
+export * from './listWebAppScans';