+++ src/steps/was/constants.ts
@@ -1,7 +1,9 @@
 import {
   generateRelationshipType,
   RelationshipClass,
+  RelationshipDirection,
   StepEntityMetadata,
+  StepMappedRelationshipMetadata,
   StepRelationshipMetadata,
 } from '@jupiterone/integration-sdk-core';
 
@@ -10,6 +12,7 @@ import { ENTITY_TYPE_SERVICE_WAS } from '../services';
 export const STEP_FETCH_SCANNED_WEBAPPS = 'fetch-scanned-webapps';
 export const STEP_FETCH_SCANNED_WEBAPP_FINDINGS =
   'fetch-scanned-webapp-findings';
+export const STEP_FETCH_ASSESSMENTS = 'fetch-assessments';
 
 export const DATA_SCANNED_WEBAPP_IDS = 'DATA_SCANNED_WEBAPP_IDS';
 export const DATA_WEBAPP_VULNERABILITY_FINDING_KEYS =
@@ -17,6 +20,7 @@ export const DATA_WEBAPP_VULNERABILITY_FINDING_KEYS =
 
 export const ENTITY_TYPE_WEBAPP = 'web_app';
 export const ENTITY_TYPE_WEBAPP_FINDING = 'qualys_web_app_finding';
+export const ENTITY_TYPE_WEBAPP_ASSESSMENT = 'qualys_web_app_assessment';
 
 export const RELATIONSHIP_TYPE_SERVICE_WEBAPP_FINDING = generateRelationshipType(
   RelationshipClass.IDENTIFIED,
@@ -30,11 +34,32 @@ export const MAPPED_RELATIONSHIP_TYPE_WAS_SCANS_WEBAPP = generateRelationshipTyp
   ENTITY_TYPE_WEBAPP,
 );
 
+export const MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_FINDING = generateRelationshipType(
+  RelationshipClass.HAS,
+  ENTITY_TYPE_WEBAPP,
+  ENTITY_TYPE_WEBAPP_FINDING,
+);
+
+export const MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_ASSESSMENT = generateRelationshipType(
+  RelationshipClass.HAS,
+  ENTITY_TYPE_WEBAPP,
+  ENTITY_TYPE_WEBAPP_ASSESSMENT,
+);
+
 export const WasEntities: Record<string, StepEntityMetadata> = {
   WEBAPP_FINDING: {
     _type: ENTITY_TYPE_WEBAPP_FINDING,
     _class: 'Finding',
     resourceName: 'Web App Finding',
+    partial: true,
+    indexMetadata: {
+      enabled: false,
+    },
+  },
+  WEBAPP_ASSESSMENT: {
+    _type: ENTITY_TYPE_WEBAPP_ASSESSMENT,
+    _class: 'Assessment',
+    resourceName: 'Web App Assessment',
     partial: true,
     indexMetadata: {
       enabled: false,
@@ -53,12 +78,41 @@ export const WasRelationships: Record<string, StepRelationshipMetadata> = {
       enabled: false,
     },
   },
+};
+
+export const WasMappedRelationships: Record<
+  string,
+  StepMappedRelationshipMetadata
+> = {
   SERVICE_SCANS_WEBAPP: {
     _type: MAPPED_RELATIONSHIP_TYPE_WAS_SCANS_WEBAPP,
     _class: RelationshipClass.SCANS,
     sourceType: ENTITY_TYPE_SERVICE_WAS,
     targetType: ENTITY_TYPE_WEBAPP,
     partial: true,
+    direction: RelationshipDirection.FORWARD,
+    indexMetadata: {
+      enabled: false,
+    },
+  },
+  WEBAPP_HAS_FINDING: {
+    _type: MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_FINDING,
+    _class: RelationshipClass.HAS,
+    sourceType: ENTITY_TYPE_WEBAPP,
+    targetType: ENTITY_TYPE_WEBAPP_FINDING,
+    partial: true,
+    direction: RelationshipDirection.FORWARD,
+    indexMetadata: {
+      enabled: false,
+    },
+  },
+  WEBAPP_HAS_ASSESSMENT: {
+    _type: MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_ASSESSMENT,
+    _class: RelationshipClass.HAS,
+    sourceType: ENTITY_TYPE_WEBAPP,
+    targetType: ENTITY_TYPE_WEBAPP_ASSESSMENT,
+    partial: true,
+    direction: RelationshipDirection.FORWARD,
     indexMetadata: {
       enabled: false,
     },