+++ src/steps/was/converters.ts
@@ -10,7 +10,10 @@ import {
   convertNumericSeverityToString,
   normalizeNumericSeverity,
 } from '../utils';
-import { ENTITY_TYPE_WEBAPP_FINDING } from './constants';
+import {
+  ENTITY_TYPE_WEBAPP_ASSESSMENT,
+  ENTITY_TYPE_WEBAPP_FINDING,
+} from './constants';
 
 export function createWebAppFindingEntity(finding: was.WebAppFinding): Entity {
   return createIntegrationEntity({
@@ -48,6 +51,34 @@ export function createWebAppFindingEntity(finding: was.WebAppFinding): Entity {
         category: 'app-scan',
         status: finding.status,
         open: !finding.status || !/fixed/i.test(finding.status),
+      },
+    },
+  });
+}
+
+export function createWebScanAssessmentEntity(
+  assessment: was.WasScanReport,
+): Entity {
+  return createIntegrationEntity({
+    entityData: {
+      source: assessment,
+      assign: {
+        _type: ENTITY_TYPE_WEBAPP_ASSESSMENT,
+        _key: assessment.TARGET.SCAN.replace(/\s+/g, '-').toLowerCase(),
+        _class: ['Assessment'],
+        name: assessment.TARGET.SCAN,
+        createdOn: parseTimePropertyValue(
+          assessment.HEADER.GENERATION_DATETIME,
+        ),
+        summary: assessment.TARGET.SCAN,
+        securityRisk: assessment.SUMMARY.GLOBAL_SUMMARY.SECURITY_RISK,
+        vulnerabilityCount: assessment.SUMMARY.GLOBAL_SUMMARY.VULNERABILITY,
+        sensitiveContentCount:
+          assessment.SUMMARY.GLOBAL_SUMMARY.SENSITIVE_CONTENT,
+        informationGatheredCount:
+          assessment.SUMMARY.GLOBAL_SUMMARY.INFORMATION_GATHERED,
+        category: 'Risk Assessment',
+        internal: true,
       },
     },
   });