+++ src/steps/was/index.ts
@@ -1,5 +1,4 @@
 import { v4 as uuid } from 'uuid';
-
 import {
   createDirectRelationship,
   createMappedRelationship,
@@ -23,8 +22,17 @@ import {
   STEP_FETCH_SCANNED_WEBAPPS,
   WasEntities,
   WasRelationships,
+  MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_FINDING,
+  STEP_FETCH_ASSESSMENTS,
+  MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_ASSESSMENT,
+  WasMappedRelationships,
 } from './constants';
-import { createWebAppFindingEntity } from './converters';
+import {
+  createWebAppFindingEntity,
+  createWebScanAssessmentEntity,
+} from './converters';
+import { SearchScanReport, WasScan } from '../../provider/client/types/was';
+import { sleep } from '../../util';
 
 export async function fetchScannedWebApps({
   logger,
@@ -96,6 +104,7 @@ export async function fetchScannedWebAppFindings({
   const serviceEntity = (await jobState.getData(
     DATA_WAS_SERVICE_ENTITY,
   )) as Entity;
+
   const vulnerabilityFindingKeysCollector = new VulnerabilityFindingKeysCollector();
 
   let numWebAppFindingsProcessed = 0;
@@ -114,6 +123,24 @@ export async function fetchScannedWebAppFindings({
           _class: RelationshipClass.IDENTIFIED,
           from: serviceEntity,
           to: findingEntity,
+        }),
+      );
+
+      await jobState.addRelationship(
+        createMappedRelationship({
+          _class: RelationshipClass.HAS,
+          _type: MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_FINDING,
+          _mapping: {
+            sourceEntityKey: findingEntity._key,
+            relationshipDirection: RelationshipDirection.REVERSE,
+            targetFilterKeys: [['_type', 'name']],
+            targetEntity: {
+              _class: 'Application',
+              _type: 'web_app',
+              name: finding.webApp?.name,
+              displayName: finding.webApp?.name,
+            },
+          },
         }),
       );
 
@@ -165,12 +192,98 @@ export async function fetchScannedWebAppFindings({
   });
 }
 
+export async function fetchAssessments({
+  logger,
+  instance,
+  jobState,
+}: IntegrationStepExecutionContext<QualysIntegrationConfig>) {
+  const apiClient = createQualysAPIClient(logger, instance.config);
+
+  const scannedWebAppIds = (await jobState.getData(
+    DATA_SCANNED_WEBAPP_IDS,
+  )) as number[];
+  const NUM_RETRIES = 5;
+
+  // Step 1 - get all the was scans
+  const wasScans: WasScan[] = [];
+  for (const webAppId of scannedWebAppIds) {
+    await apiClient.iterateWebAppScans(
+      (wasScan) => {
+        wasScans.push(wasScan);
+      },
+      {
+        filters: {
+          'webApp.id': webAppId,
+        },
+      },
+    );
+  }
+
+  // Step 2 - get all was scans IDs
+  const scanIds: number[] = [];
+  for (const wasScan of wasScans) {
+    if (wasScan.id) {
+      scanIds.push(wasScan.id);
+    }
+  }
+
+  // Step 3 - create report for each scan
+  const createdReportsIds: number[] = [];
+  for (const scanId of scanIds) {
+    const scanReportRequest = await apiClient.createScanReport(scanId);
+    createdReportsIds.push(scanReportRequest.id);
+  }
+
+  for (const createdReportId of createdReportsIds) {
+    // Step 4 - check if reports are completed
+    let reportStatus = (await apiClient.searchReport(
+      createdReportId,
+    )) as SearchScanReport;
+    let tries = 0;
+
+    while (reportStatus.status !== 'COMPLETE' && tries < NUM_RETRIES) {
+      tries++;
+      reportStatus = (await apiClient.searchReport(
+        createdReportId,
+      )) as SearchScanReport;
+      await sleep(3000);
+    }
+
+    if (reportStatus.status === 'COMPLETE') {
+      // Step 5 - get reports
+      const scanReport = await apiClient.fetchScanReport(createdReportId);
+      const assessmentEntity = await jobState.addEntity(
+        createWebScanAssessmentEntity(scanReport),
+      );
+
+      await jobState.addRelationship(
+        createMappedRelationship({
+          _class: RelationshipClass.HAS,
+          _type: MAPPED_RELATIONSHIP_TYPE_WEBAPP_HAS_ASSESSMENT,
+          _mapping: {
+            sourceEntityKey: assessmentEntity._key,
+            relationshipDirection: RelationshipDirection.REVERSE,
+            targetFilterKeys: [['_type', 'name']],
+            targetEntity: {
+              _class: 'Application',
+              _type: 'web_app',
+              name: scanReport.APPENDIX.WEBAPP.NAME,
+              displayName: scanReport.APPENDIX.WEBAPP.NAME,
+            },
+          },
+        }),
+      );
+    }
+  }
+}
+
 export const webApplicationSteps: IntegrationStep<QualysIntegrationConfig>[] = [
   {
     id: STEP_FETCH_SCANNED_WEBAPPS,
     name: 'Fetch Scanned Web Apps',
     entities: [],
-    relationships: [WasRelationships.SERVICE_SCANS_WEBAPP],
+    relationships: [],
+    mappedRelationships: [WasMappedRelationships.SERVICE_SCANS_WEBAPP],
     dependsOn: [STEP_FETCH_SERVICES],
     executionHandler: fetchScannedWebApps,
   },
@@ -180,7 +293,6 @@ export const webApplicationSteps: IntegrationStep<QualysIntegrationConfig>[] = [
     entities: [WasEntities.WEBAPP_FINDING],
     relationships: [
       WasRelationships.SERVICE_WEBAPP_FINDING,
-
       // TODO: Consider using createMappedRelationship since global mappings are
       // configured to `skipTargetCreation: true`.
 
@@ -196,7 +308,17 @@ export const webApplicationSteps: IntegrationStep<QualysIntegrationConfig>[] = [
       // `Finding.targets` containing a value that matches certain properties on
       // the `Application`.
     ],
+    mappedRelationships: [WasMappedRelationships.WEBAPP_HAS_FINDING],
     dependsOn: [STEP_FETCH_SCANNED_WEBAPPS],
     executionHandler: fetchScannedWebAppFindings,
+  },
+  {
+    id: STEP_FETCH_ASSESSMENTS,
+    name: 'Fetch Assessments',
+    entities: [WasEntities.WEBAPP_ASSESSMENT],
+    relationships: [],
+    mappedRelationships: [WasMappedRelationships.WEBAPP_HAS_ASSESSMENT],
+    dependsOn: [STEP_FETCH_SERVICES, STEP_FETCH_SCANNED_WEBAPPS],
+    executionHandler: fetchAssessments,
   },
 ];